requirements.txt
yt-dlp
torch
openai-whisper
faster-whisper
sentence-transformers
transformers
librosa
scikit-learn
numpy
opencv-python
accelerate
tqdm
requests
python-dotenv
PyYAML